[REFECTOR]: MyPage, AddCommunityPage View 분리
